requirements.txt
autopep8==1.4.4
beautifulsoup4==4.7.1
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
idna==2.8
lxml==4.3.3
numpy==1.16.3
opencv-python==4.1.0.25
pep8==1.7.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
PyMySQL==0.9.3
PyYAML==5.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
uritemplate==3.0.0
urllib3==1.25.3
wxPython==4.0.6
